requirements.txt
async-generator==1.10
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
dnspython==2.2.1
exceptiongroup==1.1.0
h11==0.14.0
idna==3.4
outcome==1.2.0
Pillow==9.4.0
pip==22.3.1
pycparser==2.21
PySocks==1.7.1
pyTelegramBotAPI==4.9.0
requests==2.28.1
selenium==4.7.2
setuptools==65.6.3
sniffio==1.3.0
sortedcontainers==2.4.0
trio-websocket==0.9.2
trio==0.22.0
urllib3==1.26.13
wheel==0.38.4
wsproto==1.2.0
